React only on snapshot from AssetManager, not delete
SchedulerService also gets notified for deletion of events, so we don't
have to react twice. Rename the handler accordingly.
